Function3
 prints the list of students with total marks less than a certain threshold. The threshold will be entered from the keyboard by the user. To
complete this function requirement, you are required to perform two steps:
o Step 1: You will create an algorithm (called Algorithm 3) in pseudo code to print the list of students with total marks less than a certain
threshold. The algorithm will take the threshold as an input.
o Step 2: You will implement Algorithm 3 based on the pseudo code developed in Step 1. Note that the implementation must be in line with
the pseudo code, and you are not allowed to use existing searching and sorting library functions in Algorithm 3.