Function4
 Prints the top 5 students with the highest total marks and the top 5 students with the lowest total marks. Like in F3, you are required to perform
two-step process to complete this function requirement:
o Step 1: You will create an algorithm (called Algorithm 4) in pseudo code to print the top 5 students with the highest total marks and the top
5 students with the lowest total marks.
o Step 2: You will implement Algorithm 4 based on the pseudo code developed in Step 1. Note that the implementation must be in line with
the pseudo code, and you are not allowed to use existing searching and sorting library functions in Algorithm 4.